Backend/requirements.txt
# Core dependencies
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1

# Google Gemini
langchain-google-genai
google-generativeai

# Article processing
newspaper3k==0.2.8
beautifulsoup4==4.13.3
lxml==5.3.1
tldextract==5.1.3
nltk==3.9.1

# Web/API utilities
requests==2.32.3
feedparser==6.0.11

# Pydantic (for FastAPI models)
pydantic==2.10.6

# Optional but useful
tenacity  # For retry logic


# Rate limiting
slowapi==0.1.8